crypto-dashboard/requirements.txt
streamlit==1.28.1
pyspark==3.5.0
pandas==2.0.3
requests==2.31.0
psycopg2-binary==2.9.7
seaborn==0.12.2
matplotlib==3.7.2
python-dotenv==1.0.0
